docker/authService/requirements.txt
flask
flask_restful
mysql-connector
kubernetes